+++ src/actions/DashboardPageActions.js
@@ -1,6 +1,7 @@
 import ActionTypes from '../constants/ActionTypes';
 import DashboardBalancesService from '../services/DashboardBalancesService';
 import Repository from '../repositories/chain/repository';
+import GposService from '../services/GposService';
 
 class DashboardPagePrivateActions {
   /**
@@ -87,6 +88,33 @@ class DashboardPagePrivateActions {
   static setVestingBalancesAction(data) {
     return {
       type: ActionTypes.DASHBOARD_SET_SIDE_VESTING_BALANCES,
+      payload: data
+    };
+  }
+
+  /**
+   * Private Redux Action Creator
+   * GPOS Balances
+   * @param {data} data
+   * @returns {{type, payload: Immutable.Map()}}
+   */
+  static setGposBalancesAction(data) {
+    return {
+      type: ActionTypes.DASHBOARD_SET_GPOS_BALANCES,
+      payload: data
+    };
+  }
+
+  /**
+   * Private Redux Action Creator
+   * GPOS Info
+   *
+   * @param {object} data
+   * @memberof DashboardPagePrivateActions
+   */
+  static setGposInfoAction(data) {
+    return {
+      type: ActionTypes.DAHSBOARD_SET_GPOS_INFO,
       payload: data
     };
   }
@@ -181,8 +209,13 @@ class DashboardPageActions {
         vestingAsset = currentState.dashboardPage.vestingAsset,
         vestingBalancesIds = currentState.dashboardPage.vestingBalancesIds;
 
+      // GPOS
+      let gposBalances = currentState.dashboardPage.gposBalances;
+      this.getGposInfo();
+
+      // Get regular vesting balances
       DashboardBalancesService
-        .calculateVesting(currentState.app.account, vestingBalances).then((data) => {
+        .calculateVesting(currentState.app.account, vestingBalances, gposBalances).then((data) => {
           if (!data) {
             return null;
           }
@@ -198,7 +231,43 @@ class DashboardPageActions {
               vestingAsset: data.vestingAsset
             }));
           }
+
+          if (gposBalances !== data.gposBalances) {
+            dispatch(DashboardPagePrivateActions.setGposBalancesAction({
+              gposBalances: data.gposBalances
+            }));
+          }
         });
+    };
+  }
+
+  /**
+   * Retrieve the GPOS information for the logged in user.
+   *
+   * @static
+   * @returns
+   * @memberof DashboardPageActions
+   */
+  static getGposInfo() {
+    return (dispatch, getState) => {
+      let currentState = getState();
+      GposService.fetchGposInfo(currentState.app.accountId).then((info) => {
+        dispatch(DashboardPageActions.setGposInfo(info));
+      });
+    };
+  }
+
+  /**
+   * Set the GPOS information to the redux state store.
+   *
+   * @static
+   * @param {object} info
+   * @returns
+   * @memberof DashboardPageActions
+   */
+  static setGposInfo(info) {
+    return (dispatch) => {
+      dispatch(DashboardPagePrivateActions.setGposInfoAction(info));
     };
   }