+++ src/actions/RSettingsActions.js
@@ -12,7 +12,7 @@ let ss = new ls(STORAGE_KEY);
 let storageSettings = ss.get(SETTINGS_KEY);
 let locales = {};
 ['cn', 'de', 'es', 'fr', 'ko', 'tr'].forEach((locale) => {
-  locales[locale] = require('json!../assets/locales/locale-' + locale + '.json');
+  locales[locale] = require('../assets/locales/locale-' + locale + '.json');
 });
 
 // export function
@@ -39,6 +39,7 @@ class RSettingsAcions {
         let newSettings = {};
         let needUpdate = false;
 
+        // eslint-disable-next-line no-unused-vars
         for (let key in storageSettings) {
           if (storageSettings[key] !== state.settings[key]) {
             if (key === 'hiddenAssets') {