+++ src/assets/loader.js
@@ -1,7 +1,5 @@
 /* eslint-disable */
-require('file?name=index.html!../../public/index.html');
-require('file?name=favicon.ico!./favicon.ico');
-require('file?name=dictionary.json!../common/dictionary_en.json');
+require('file-loader?name=dictionary.json!../common/dictionary_en.json');
 require('babel-polyfill');
 require('whatwg-fetch');
 require('indexeddbshim');