+++ src/assets/styling/rules/declaration.js
@@ -23,14 +23,15 @@ module.exports = {
   ],
   'declaration-no-important': true,
   'declaration-property-unit-whitelist': {
-    '/^animation/': ['s'],
+    '/^animation/': ['s', 'ms'],
+    'animation': ['ms'],
     'background': ['px', 'x', '%'],
     'background-image': ['px', 'x', '%'],
     'border': ['px'],
-    'font-size': ['px'],
+    'font-size': ['px', 'rem'],
     'font': ['px'],
     'height': ['px', '%', 'vw', 'vh'],
-    'line-height': [],
+    'line-height': ['px'],
     'margin': ['px'],
     'margin-top': ['px'],
     'margin-right': ['px'],
@@ -41,6 +42,6 @@ module.exports = {
     'padding-right': ['px'],
     'padding-bottom': ['px'],
     'padding-left': ['px'],
-    'width': ['px', '%', 'vw', 'vh'],
+    'width': ['px', '%', 'vw', 'vh']
   }
 };