+++ src/common/localStorage.js
@@ -5,7 +5,7 @@ if (null === ls) {
   throw 'localStorage is required but isn\'t available on this platform';
 }
 
-module.exports = (key) => {
+export default (key) => {
   var STORAGE_KEY = key;
 
   return {