+++ src/selectors/SideVesting.js
@@ -10,14 +10,20 @@ export const getTotalVestingBalances = createSelector(
       totalClaimable = 0;
 
     balanceIds.forEach((balanceId) => {
-      let balance = balances.get(balanceId),
-        balanceAmount = balance.getIn(['balance', 'amount']),
-        vestingPeriod = balance.getIn(['policy', 1, 'vesting_seconds']),
-        earned = balance.getIn(['policy', 1, 'coin_seconds_earned']),
-        availablePercent = earned / (vestingPeriod * balanceAmount),
-        claim = balanceAmount * availablePercent;
-      totalAmount += balanceAmount;
-      totalClaimable += claim;
+      if (balances.size > 0) {
+        let balance = balances.get(balanceId);
+
+        if (balance !== undefined) {
+          let balanceAmount = balance.getIn(['balance', 'amount']),
+            vestingPeriod = balance.getIn(['policy', 1, 'vesting_seconds']),
+            earned = balance.getIn(['policy', 1, 'coin_seconds_earned']),
+            availablePercent = earned / (vestingPeriod * balanceAmount),
+            claim = balanceAmount * availablePercent;
+
+          totalAmount += balanceAmount;
+          totalClaimable += claim;
+        }
+      }
     });
 
     return {