+++ src/services/DashboardBalancesService.js
@@ -148,26 +148,6 @@ class DashboardBalancesService {
               }
             });
 
-            // let endDate = new Date();
-            // let startDateShort = new Date();
-            //
-            // let baseAsset = assets.find((asset) => {
-            // 	return asset.id === '1.3.0';
-            // });
-            //
-            // endDate.setDate(endDate.getDate() + 1);
-            // startDateShort = new Date(startDateShort.getTime() - 3600 * 50 * 1000);
-
-            // let promisesStats = [];
-
-            // assets.forEach((asset) => {
-            // 	promisesStats.push(
-            // self.calcStats(asset, baseAsset, startDateShort.toISOString().slice(0, -5),
-            // endDate.toISOString().slice(0, -5)));
-            // });
-            //
-            // return Promise.all(promisesStats).then(() => {
-
             cacheData.assets = assets;
             cacheData.amounts = amounts;
             cacheData.marketStatsByName = marketStatsByName;
@@ -183,11 +163,8 @@ class DashboardBalancesService {
             dataIsFetched = true;
 
             return true;
-
-            // });
           });
         });
-
       });
     });
   }
@@ -195,9 +172,10 @@ class DashboardBalancesService {
   /**
    *
    * @param {string} accountId
-   * @param {Immutable.Map} vestingBalances
+   * @param {Immutable.Map} vestingBalances,
+   * @param {string} type - regular or GPOS. Default to witness vest type.
    */
-  static calculateVesting(accountId, vestingBalances) {
+  static calculateVesting(accountId, vestingBalances, gposBalances) {
     return Repository.fetchFullAccount(accountId).then((account) => {
       if (!account) {
         return null;
@@ -214,18 +192,25 @@ class DashboardBalancesService {
         .then(([balances, asset]) => {
           if (balances && balances.length) {
             balances.forEach((balance) => {
-
-              if (balance) {
+              // Only work with non-gpos vested balances.
+              if (balance && balance.get('balance_type').toLowerCase() !== 'gpos') {
                 let currBalance = vestingBalances.get(balance.get('id'));
 
                 if (!currBalance || currBalance !== balance) {
                   vestingBalances = vestingBalances.set(balance.get('id'), balance);
+                }
+              } else {
+                let currGposBalance = gposBalances.get(balance.get('id'));
+
+                if (!currGposBalance || currGposBalance !== balance) {
+                  gposBalances = gposBalances.set(balance.get('id'), balance);
                 }
               }
             });
           }
 
           return {
+            gposBalances,
             vestingBalancesIds: vesting_balances_ids ? vesting_balances_ids : [],
             vestingBalances: vestingBalances,
             vestingAsset: asset
@@ -504,22 +489,6 @@ class DashboardBalancesService {
       latestPrice
     };
   }
-
-  // static calcStats(asset, baseAsset, start, end) {
-  // 	return Promise.all([
-  // 		HistoryRepository.fetchMarketHistory(baseAsset.id, asset.id, 3600, start, end),
-  // 		HistoryRepository.fetchFillOrderHistory(baseAsset.id, asset.id, 1)
-  // 	]).then(result => {
-  //
-  // 		let history = result[0],
-  // 			last = result[1],
-  // 			stats = this._calcMarketStats(history, baseAsset, asset, last),
-  // 			marketName = asset.symbol + "_" + baseAsset.symbol;
-  //
-  // 		marketStatsByName[marketName] = stats;
-  //
-  // 	});
-  // }
 
   /**
    *